+++ track/src/main/resources/db/migration/V11__init_session_gelernt_event_table.sql
@@ -8,7 +8,7 @@ create table session_beendet_event(
 
 create table session_bewertung(
     session_beendet_event int references session_beendet_event(id) on delete cascade,
-    konzentration_bewertung decimal check (konzentration_bewertung between 0 and 10),
-    produktivitaet_bewertung decimal check (produktivitaet_bewertung between 0 and 10),
-    schwierigkeit_bewertung decimal check (schwierigkeit_bewertung between 0 and 10)
+    konzentration_bewertung int check (konzentration_bewertung between 0 and 10),
+    produktivitaet_bewertung int check (produktivitaet_bewertung between 0 and 10),
+    schwierigkeit_bewertung int check (schwierigkeit_bewertung between 0 and 10)
 );
